Add volume_leader_exclusions table and use it to filter query (#524)
It's possible for our fee payers to show up in the leaderboard. While
they wouldn't win anything, it hurts the experience. Adding a table that
can be used to filter out records from showing in the endpoint.
I manually ran the migration and added a list of addresses with
descriptions to the table.