Fix duplicate rows in /coins if multiple reward pool accounts (#577)
- I accidentally created a non-ATA account on the authority of PHUTURES
- Later created the ATA
- `/v1/coins` was giving duplicate rows for PHUTURES
- Fix by only selecting earliest-created account for that owner - as a
proxy for the ATA, assuming we'd create it as part of the reward pool
creation process.
- But technically the PHUTURES response is still showing the wrong
account (since I created the non-ATA before the ATA). Not a huge issue
bc we don't use that value anywhere. But maybe I can just delete that
row for this case?

Tested on prod