Add UploadNodes to control uploads more tightly (#559)
Creates a configurable list of nodes that will be used for uploads (as
they are returned by health_check)

Keeps existing health monitoring for all nodes, just filters at the end.
A "smarter" method might be filtering _before_ tracking health, but this
gets the job done for now. Opted to do the filtering outside of the
monitor as it might be confusing why it only has a small number of nodes
it thinks are "healthy".

Notably still might be a footgun, as it's not obvious that health check
would be limited to the UploadNodes